auth in soap and paging

+++ app/src/main/java/github/sachin2dehury/owlmail/data/auth/AuthResponse.kt
@@ -10,5 +10,5 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class AuthResponse(
     val authToken: List<Content>? = null,
-    val lifetime: Int? = null
+    val lifetime: Long? = null
 ) : Parcelable
